Give every corpus case an oracle, and settle switch-fold on evidence
Two items, both about the harnesses telling the truth.

## Three cases were comparing mtsc against itself

`verify_mangle_safety` had been printing "reference run unavailable" on
every run for case09-complex, case10-module and case18-export-type.
Without the reference leg those three are baseline-vs-mangled only — a
self-comparison, which is consistency and not correctness, and which is
the exact failure mode that let five scope-narrowing bugs survive
thousands of fuzz seeds. It was sitting inside the corpus everything
else leans on.

Each was one fixture line, and each was a COMPILATION question wearing a
mangle-safety costume:

  case18  `import { LocalObj }` — a VALUE-form import of a type.
          TypeScript accepts it and mtsc erases it; Node's transform
          mode keeps the import and `./types` has no runtime export.
  case10  the legacy `module X {}` keyword, which transform mode
          rejects outright. The case already covered `namespace`
          separately in the very next block, so the keyword was the
          only unique thing in it.
  case09  a TypeScript-legacy `@log` property decorator — the "Invalid
          or unexpected token" — bundled into a grab-bag case with
          generics, enums, namespaces and mapped types, all of which
          lost their oracle with it.

The first two move to unit tests, where a compilation question belongs:
one asserts the value-form type import leaves no runtime import, the
other that `module X {}` and `namespace X {}` emit byte-identical JS.
The decorator is split into `case09b-decorator`, which declares
`reference: false` with the reason, so the rest of case09 now runs
against the original TypeScript.

174 cases pass and the harness prints no "reference run unavailable"
notes at all. Every case has an independent oracle except the one that
says why it cannot.

## switch-fold: kept, unchanged, and recorded

The pass moves zero bytes on the whole corpus and the shape is absent —
seven of nine targets have no candidate. The report now names the three
candidates in the other two, because a count never says what to look
at, and every one is a false positive of the entry test:

  typebox     `LiteralBooleanMapping`, `WithBooleanMapping` — bodies are
              `return T.Literal(Guard.IsEqual(input, 'true'))`, no
              `switch` at all, and both call sites pass
              `_0 as 'true' | 'false'`. Widening the BODY gate cannot
              help: the call-site gate needs a `StringLit`.
  excalidraw  a class `constructor`, which no widening can replace with
              an arm expression.

So there is nothing to reach — and deleting the pass is equally wrong.
It is the only thing that wins the terser-parity `switch-literal-union`
case, which is a `type-aware` case where a TIE already counts as a
failure, and its one historical bug (the name-keyed table taking the
outer dispatcher's arm for a call to a shadowing parameter, a wrong
value rather than a crash) is covered under Node by
`case43-table-shadowing` alongside an unshadowed read.

Kept as is, with the measured reach and both reasons written into the
pass header so the question is not re-litigated.

Verified: 2905 tests, 174 corpus cases with zero missing oracles, 31 DCE
eliminated + 1 labelled miss, 32 win / 0 loss vs terser, real-world
byte-identical (TypeScript 9,044,103 -> 3,585,914).
