+++ fixtures/mangle-safety/case09b-decorator/driver.mjs
@@ -1,0 +1,3 @@
+export default async (mod) => ({
+  decorated: new mod.DecoratorTest(),
+});
